chore: reorganize DE cell into new dir
