+++ src/data-workspace/data-entry-cell/final-form-wrapper.js
@@ -1,0 +1,20 @@
+import PropTypes from 'prop-types'
+import React from 'react'
+import { Form } from 'react-final-form'
+
+export function FinalFormWrapper({ children, initialValues }) {
+    return (
+        <Form
+            onSubmit={(values, form) => {
+                console.log({ values, form })
+            }}
+            initialValues={initialValues}
+        >
+            {() => children}
+        </Form>
+    )
+}
+FinalFormWrapper.propTypes = {
+    children: PropTypes.node,
+    initialValues: PropTypes.any,
+}
