+++ src/data-workspace/data-entry-cell/index.js
@@ -1,0 +1,2 @@
+export { FinalFormWrapper } from './final-form-wrapper.js'
+export { DataEntryCell } from './data-entry-cell.js'
